backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.1.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
